# coffee-be/static/swagger.json
{
	"components": {
...
		"schemas": {
...
			"ChainActionType": {
				"description": "Codebook for action type",
				"properties": {
					"docType": {
						"type": "string",
						"nullable": true
					},
					"_id": {
						"type": "string",
						"nullable": true
					},
					"_rev": {
						"type": "string",
						"nullable": true
					},
					"id": {
						"type": "string"
					},
					"label": {
						"type": "string"
					},
					"facilityId": {
						"type": "string"
					},
					"facilityType": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"label"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ChainFacilityType": {
				"description": "Codebook for type of chain facility",
				"properties": {
					"docType": {
						"type": "string",
						"nullable": true
					},
					"_id": {
						"type": "string",
						"nullable": true
					},
					"_rev": {
						"type": "string",
						"nullable": true
					},
					"id": {
						"type": "string"
					},
					"label": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"label"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ApiResponse_any_": {
				"properties": {
					"data": {
						"additionalProperties": true
					},
					"errorDetails": {
						"type": "string",
						"description": "Optional details for unexpected error responses."
					},
					"errorMessage": {
						"type": "string",
						"description": "Simple message to explain client developers the reason for error."
					},
					"status": {
						"$ref": "#/components/schemas/ApiDefaultResponseStatusEnum",
						"description": "Response status. OK for successful reponses."
					},
					"validationErrorDetails": {
						"$ref": "#/components/schemas/ApiValidationErrorDetails"
					}
				},
				"required": [
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ApiResponse_PaginatedList_ChainActionType__": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/PaginatedList_ChainActionType_"
					},
					"errorDetails": {
						"type": "string",
						"description": "Optional details for unexpected error responses."
					},
					"errorMessage": {
						"type": "string",
						"description": "Simple message to explain client developers the reason for error."
					},
					"status": {
						"$ref": "#/components/schemas/ApiDefaultResponseStatusEnum",
						"description": "Response status. OK for successful reponses."
					},
					"validationErrorDetails": {
						"$ref": "#/components/schemas/ApiValidationErrorDetails"
					}
				},
				"required": [
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_ChainActionType_": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/ChainActionType"
					},
					"errorDetails": {
						"type": "string",
						"description": "Optional details for unexpected error responses."
					},
					"errorMessage": {
						"type": "string",
						"description": "Simple message to explain client developers the reason for error."
					},
					"status": {
						"$ref": "#/components/schemas/ApiDefaultResponseStatusEnum",
						"description": "Response status. OK for successful reponses."
					},
					"validationErrorDetails": {
						"$ref": "#/components/schemas/ApiValidationErrorDetails"
					}
				},
				"required": [
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "coffee-be",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/chain-api/codebook/action-type/list": {
			"get": {
				"operationId": "GetActionTypeList",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_PaginatedList_ChainActionType__"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Returns paginated list of action types.",
				"tags": [
					"Codebook"
				],
				"security": [],
				"parameters": [
					{
						"description": "sort order ASC or DESC",
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {}
					},
					{
						"description": "query limit",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "query offset",
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/chain-api/codebook/action-type/{dbId}": {
			"get": {
				"operationId": "GetActionType",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_ChainActionType_"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Fetches an action type by its _id.",
				"tags": [
					"Codebook"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "dbId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/chain-api/codebook/action-type": {
			"post": {
				"operationId": "PostActionType",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Inserts or updates an action type. When inserting fields _id, _rev and docType should not be present.",
				"tags": [
					"Codebook"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChainActionType"
							}
						}
					}
				}
			}
		},
		"/chain-api/codebook/action-type/delete": {
			"post": {
				"operationId": "DeleteActionType",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Deletes action type.",
				"tags": [
					"Codebook"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChainActionType"
							}
						}
					}
				}
			}
		},
		"/chain-api/codebook/facility-type/list": {
			"get": {
				"operationId": "GetFacilityTypeList",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_PaginatedList_ChainFacilityType__"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Returns paginated list of facility types.",
				"tags": [
					"Codebook"
				],
				"security": [],
				"parameters": [
					{
						"description": "sort order ASC or DESC",
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {}
					},
					{
						"description": "query limit",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "query offset",
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/chain-api/codebook/facility-type/{dbId}": {
			"get": {
				"operationId": "GetFacilityType",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_ChainFacilityType_"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Fetches a facility type by its _id.",
				"tags": [
					"Codebook"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "dbId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/chain-api/codebook/facility-type": {
			"post": {
				"operationId": "PostFacilityType",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Inserts or updates a facility type. When inserting fields _id, _rev and docType should not be present.",
				"tags": [
					"Codebook"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChainFacilityType"
							}
						}
					}
				}
			}
		},
		"/chain-api/codebook/facility-type/delete": {
			"post": {
				"operationId": "DeleteFacilityType",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Deletes facility type.",
				"tags": [
					"Codebook"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChainFacilityType"
							}
						}
					}
				}
			}
		},
		"/chain-api/codebook/measure-unit-type/list": {
			"get": {
				"operationId": "GetMeasureUnitTypeList",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_PaginatedList_ChainMeasureUnitType__"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Returns paginated list of measure unit types.",
				"tags": [
					"Codebook"
				],
				"security": [],
				"parameters": [
					{
						"description": "sort order ASC or DESC",
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {}
					},
					{
						"description": "query limit",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "query offset",
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/chain-api/codebook/measure-unit-type/{dbId}": {
			"get": {
				"operationId": "GetMeasureUnitType",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_ChainMeasureUnitType_"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Fetches a measure unit type by its _id.",
				"tags": [
					"Codebook"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "dbId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/chain-api/codebook/measure-unit-type": {
			"post": {
				"operationId": "PostMeasureUnitType",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Inserts or updates a measure unit type. When inserting fields _id, _rev and docType should not be present.",
				"tags": [
					"Codebook"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChainMeasureUnitType"
							}
						}
					}
				}
			}
		},
		"/chain-api/codebook/measure-unit-type/delete": {
			"post": {
				"operationId": "DeleteMeasureUnitType",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Deletes measure unit type.",
				"tags": [
					"Codebook"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChainMeasureUnitType"
							}
						}
					}
				}
			}
		},
		"/chain-api/codebook/grade-abbreviation/list": {
			"get": {
				"operationId": "GetGradeAbbreviationList",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_PaginatedList_ChainGradeAbbreviation__"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Returns paginated list of grade abbreviations.",
				"tags": [
					"Codebook"
				],
				"security": [],
				"parameters": [
					{
						"description": "sort order ASC or DESC",
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {}
					},
					{
						"description": "query limit",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "query offset",
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/chain-api/codebook/grade-abbreviation/{dbId}": {
			"get": {
				"operationId": "GetGradeAbbreviation",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_ChainGradeAbbreviation_"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Fetches a grade abbreviation by its _id.",
				"tags": [
					"Codebook"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "dbId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/chain-api/codebook/grade-abbreviation": {
			"post": {
				"operationId": "PostGradeAbbreviation",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Inserts or updates a grade abbreviation. When inserting fields _id, _rev and docType should not be present.",
				"tags": [
					"Codebook"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChainGradeAbbreviation"
							}
						}
					}
				}
			}
		},
		"/chain-api/codebook/grade-abbreviation/delete": {
			"post": {
				"operationId": "DeleteGradeAbbreviation",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Deletes grade abbreviation.",
				"tags": [
					"Codebook"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChainGradeAbbreviation"
							}
						}
					}
				}
			}
		},
		"/chain-api/codebook/processing-evidence-type/list": {
			"get": {
				"operationId": "GetProcessingEvidenceTypeList",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_PaginatedList_ChainProcessingEvidenceType__"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Returns paginated list of processing evidence type.",
				"tags": [
					"Codebook"
				],
				"security": [],
				"parameters": [
					{
						"description": "sort order ASC or DESC",
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {}
					},
					{
						"description": "query limit",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "query offset",
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/chain-api/codebook/processing-evidence-type/{dbId}": {
			"get": {
				"operationId": "GetProcessingEvidenceType",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_ChainProcessingEvidenceType_"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Fetches a processing evidence type by its _id.",
				"tags": [
					"Codebook"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "dbId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/chain-api/codebook/processing-evidence-type": {
			"post": {
				"operationId": "PostProcessingEvidenceType",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Inserts or updates a rocessing evidence type. When inserting fields _id, _rev and docType should not be present.",
				"tags": [
					"Codebook"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChainProcessingEvidenceType"
							}
						}
					}
				}
			}
		},
		"/chain-api/codebook/processing-evidence-type/delete": {
			"post": {
				"operationId": "DeleteProcessingEvidenceType",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Deletes rocessing evidence type.",
				"tags": [
					"Codebook"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChainProcessingEvidenceType"
							}
						}
					}
				}
			}
		},
		"/chain-api/codebook/order-evidence-type/list": {
			"get": {
				"operationId": "GetOrderEvidenceTypeList",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_PaginatedList_ChainOrderEvidenceType__"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Returns paginated list of order evidence type.",
				"tags": [
					"Codebook"
				],
				"security": [],
				"parameters": [
					{
						"description": "sort order ASC or DESC",
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {}
					},
					{
						"description": "query limit",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "query offset",
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/chain-api/codebook/order-evidence-type/{dbId}": {
			"get": {
				"operationId": "GetOrderEvidenceType",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_ChainOrderEvidenceType_"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Fetches a order evidence type by its _id.",
				"tags": [
					"Codebook"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "dbId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/chain-api/codebook/order-evidence-type": {
			"post": {
				"operationId": "PostOrderEvidenceType",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Inserts or updates a order evidence type. When inserting fields _id, _rev and docType should not be present.",
				"tags": [
					"Codebook"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChainOrderEvidenceType"
							}
						}
					}
				}
			}
		},
		"/chain-api/codebook/order-evidence-type/delete": {
			"post": {
				"operationId": "DeleteOrderEvidenceType",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Deletes order evidence type.",
				"tags": [
					"Codebook"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChainOrderEvidenceType"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}